feat: Create Config, expanded Serializable
1. Config class
Config class is used to store values with keys.
It works similar to dictionary, but it only accepts string as key.
So, Config objects can be serialized into JSON.

2. Serializable, BytesSerializable, JSONSerializable
Serializable is abstarct class of all Serializable objects.
It defines serialization/deserialization methods in it.
BytesSerializable and JSONSerializable is a subclass of Serializable.
They specify serialized data's types.